* Killed "phase" terminology. * Updated progress tracking to not need "phases". * Reviewed by Micah.

+++ trunk/scipolicy/plugins/normal_plugins/edu.iu.scipolicy.database.isi.load/src/edu/iu/scipolicy/database/isi/load/utilities/parser/ISITableModelParser.java
@@ -154,7 +154,7 @@ public class ISITableModelParser {
 
 		int last = 0;
 		int total = 0;
-		int unitsWorked = 0;
+		double unitsWorked = 0;
 
 		for (Integer rowIndex : rows) {
 			ProgressMonitorUtilities.handleCanceledOrPausedAlgorithm(progressMonitor);
